documented the cane macro + made few improvements
