+++ src/main/java/me/kokoniara/kokoMod/kokoMod.java
@@ -1,7 +1,6 @@
 package me.kokoniara.kokoMod;
 
 import me.kokoniara.kokoMod.config.configMenager;
-import me.kokoniara.kokoMod.renderEngine.renderEngine;
 import org.lwjgl.input.Keyboard;
 
 import me.kokoniara.kokoMod.clickgui.ClickGui;
@@ -20,7 +19,6 @@ public class kokoMod
     public SettingsManager settingsManager;
     public ClickGui clickGui;
     public configMenager configMenagerINSTANCE;
-    public renderEngine renderEngine;
 
     
     public void init() {
@@ -28,7 +26,6 @@ public class kokoMod
     	settingsManager = new SettingsManager();
     	moduleManager = new ModuleManager();
     	clickGui = new ClickGui();
-        renderEngine = new renderEngine();
     }
     
     @SubscribeEvent
