+++ src/main/java/me/kokoniara/kokoMod/module/misc/amiTimedOut.java
@@ -18,7 +18,4 @@ public class amiTimedOut extends Module {
         sendChatMessage.sendClientMessage("serverPing: " + serverPing, false);
     }
 
-
-
-
 }
