+++ src/main/java/me/kokoniara/kokoMod/module/misc/macos/macroStages.java
@@ -1,4 +1,4 @@
-package me.kokoniara.kokoMod.module.misc;
+package me.kokoniara.kokoMod.module.misc.macos;
 
 public enum macroStages {
     LEFT,RIGHT,TOP,BOTTOM,HOLD,DEFAULT
