+++ src/main/java/me/kokoniara/kokoMod/module/player/iWillcancelYouOnTwitter.java
@@ -3,10 +3,8 @@ package me.kokoniara.kokoMod.module.player;
 import me.kokoniara.kokoMod.kokoMod;
 import me.kokoniara.kokoMod.module.Category;
 import me.kokoniara.kokoMod.module.Module;
-import me.kokoniara.kokoMod.renderEngine.modules.renderchartColumn;
 import me.kokoniara.kokoMod.settings.Setting;
 import me.kokoniara.kokoMod.util.sendChatMessage;
-import net.minecraftforge.client.event.RenderWorldLastEvent;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 import net.minecraftforge.fml.common.gameevent.TickEvent;
 
@@ -14,7 +12,7 @@ public class iWillcancelYouOnTwitter extends Module {
     private long lastSpam;
     private double speed;
     public iWillcancelYouOnTwitter(){
-        super("twitterWhiteGirls", "white twitter girls are the downfal on sociaty", Category.PLAYER, true, "iwillcancelyou enabled", "iwillcancelyou disabled");
+        super("twitterWhiteGirls", "white twitter girls are the downfall on society", Category.PLAYER, true, " iwillcancelyou enabled", " iwillcancelyou disabled");
         kokoMod.instance.settingsManager.rSetting(new Setting("cancel speed", this, 12, 1, 20, true));
     }
 
@@ -25,10 +23,6 @@ public class iWillcancelYouOnTwitter extends Module {
             lastSpam = System.currentTimeMillis();
 
         }
-    }
-    @SubscribeEvent
-    public void RenderLivingEvent(RenderWorldLastEvent event){
-        renderchartColumn.renderWeirdQuad(event);
     }
 
     @Override
