--- src/main/java/me/kokoniara/kokoMod/renderEngine/renderEngine.java
@@ -1,8 +1,0 @@
-package me.kokoniara.kokoMod.renderEngine;
-
-import me.kokoniara.kokoMod.renderEngine.modules.drawCenterString;
-
-
-public class renderEngine {
-    public drawCenterString drawCenterString = new drawCenterString();
-}
